Fix tilde expansion tests for environments with non-existent $HOME.
Fixes: #238

+++ t/fits.t
@@ -3,6 +3,8 @@
 
 use strict;
 
+use File::Basename;
+
 use PDL::LiteF;
 
 use PDL::Core ':Internal'; # For howbig()
@@ -319,9 +321,11 @@ SKIP:{
 ###############################
 # Check that tilde expansion works
 my $tildefile = cfile('~',"PDL-IO-FITS-test_$$.fits");
-lives_ok { sequence(3,5,2)->wfits($tildefile) } "wfits tilde expansion didn't fail";
-lives_ok { rfits($tildefile) } "rfits tilde expansion didn't fail";
+lives_ok { if(-w dirname($tildefile)) { sequence(3,5,2)->wfits($tildefile) } } "wfits tilde expansion didn't fail";
+lives_ok { if(-w dirname($tildefile)) { rfits($tildefile) } } "rfits tilde expansion didn't fail";
 $tildefile =~ s/^(~)/glob($1)/e; #use the same trick as in FITS.pm to resolve this filename.
-unlink($tildefile) or warn "Could not delete $tildefile: $!\n"; #clean up.
+if(-e $tildefile) {
+	unlink($tildefile) or warn "Could not delete $tildefile: $!\n"; #clean up.
+}
 
 1;
